# docs/04-Backend/archive/v3-openAPI.yaml
openapi: "3.1.0" 
info:
  title: "MWAP API"
  version: "1.0.0"
  description: "Full OpenAPI spec for the MWAP backend aligned with the domain model, contract, and architecture."
servers:
  - url: "https://api.mwap.dev/api/v1"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Tenant:
      type: object
      properties:
        _id: { type: string }
        ownerId: { type: string }
        name: { type: string }
        createdAt: { type: string, format: date-time }
        archived: { type: boolean }
        integrations:
          type: array
          items: { type: string }
...
paths:
...
  /tenants/me:
    get:
      summary: Get current tenant
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
...
  /tenants/{tenantId}/integrations:
    get:
      summary: List tenant integrations
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { type: object }

    post:
      summary: Create tenant integration
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: object }
...
  /tenants:
    get:
      summary: List all tenants (superadmin only)
      description: Retrieves a list of all tenants. Requires superadmin role.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tenant'
  /tenants/{id}:
    get:
      summary: Get tenant by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
...